+++ Lib/fontbakery/profiles/opentype.py
@@ -65,7 +65,8 @@ OPENTYPE_PROFILE_CHECKS = [
     'com.google.fonts/check/points_out_of_bounds',
     'com.google.fonts/check/all_glyphs_have_codepoints',
     'com.google.fonts/check/monospace_max_advancewidth',
-    'com.google.fonts/check/wght_valid_range'
+    'com.google.fonts/check/wght_valid_range',
+    'com.google.fonts/check/code_pages',
 ]
 
 profile.auto_register(globals())